+++ R/basic_example.R
@@ -4,7 +4,7 @@
 #' Also writes/overwrites the file \code{report.Rmd}.
 #' For a thorough walkthrough of how to set up this example, see the
 #' quickstart vignette: \code{vignette('quickstart')}. Alternatively,
-#' call \code{\link{example_drake}('basic')} to generate an R script
+#' call \code{\link{drake_example}('basic')} to generate an R script
 #' that builds up this example step by step.
 #' @export
 #' @return The workflow plan data frame of the basic example.