+++ R/read.R
@@ -153,7 +153,7 @@ load_target <- function(target, cache, envir, verbose){
   assign(x = target, value = value, envir = envir)
 }
 
-#' @title Function \code{read_config}
+#' @title Function \code{read_drake_config}
 #' @description Read all the configuration parameters
 #' from your last attempted call to \code{\link{make}()}.
 #' These include the workflow plan
@@ -176,9 +176,9 @@ load_target <- function(target, cache, envir, verbose){
 #' load_basic_example() # Load drake's canonical example.
 #' make(my_plan) # Run the project, build the targets.
 #' # Retrieve the master internal configuration list from the cache.
-#' read_config()
+#' read_drake_config()
 #' }
-read_config <- function(path = getwd(), search = TRUE,
+read_drake_config <- function(path = getwd(), search = TRUE,
   cache = drake::get_cache(path = path, search = search, verbose = verbose),
   verbose = TRUE
 ){
@@ -222,10 +222,10 @@ read_plan <- function(path = getwd(), search = TRUE,
   cache = drake::get_cache(path = path, search = search, verbose = verbose),
   verbose = TRUE
 ){
-  read_config(path = path, search = search, cache = cache)$plan
+  read_drake_config(path = path, search = search, cache = cache)$plan
 }
 
-#' @title Function \code{read_graph}
+#' @title Function \code{read_drake_graph}
 #' @description Read the igraph-style dependency graph of your targets
 #' from your last attempted call to \code{\link{make}()}.
 #' For better graphing utilities, see \code{\link{vis_drake_graph}()}
@@ -251,14 +251,14 @@ read_plan <- function(path = getwd(), search = TRUE,
 #' load_basic_example() # Load the canonical example.
 #' make(my_plan) # Run the project, build the targets.
 #' # Retrieve the igraph network from the cache.
-#' g <- read_graph()
+#' g <- read_drake_graph()
 #' class(g) # "igraph"
 #' }
-read_graph <- function(path = getwd(), search = TRUE,
+read_drake_graph <- function(path = getwd(), search = TRUE,
   cache = drake::get_cache(path = path, search = search, verbose = verbose),
   verbose = TRUE,
   ...
 ){
-  config <- read_config(path = path, search = search, cache = cache)
+  config <- read_drake_config(path = path, search = search, cache = cache)
   return(config$graph)
 }